inyector/requirements.txt
virtualenv==20.13.0
fastapi==0.89.1
pydantic==1.10.4
uvicorn==0.20.0
kafka-python==2.0.2